requirements.txt
pycryptodomex==3.19.1
protobuf>=5.27.2,<6.0.0
cryptography==43.0.1
pynacl==1.5.0
mnemonic==0.20
requests>=2.32.0,<3.0.0
robotframework==7.1.1
robotframework-tidy==4.14.0